Properly handle contextualization with multiprocessing.
Formulas, atoms, and lemmas must be contextualized first in the workers' environments.
Workers' lemmas must be then contextualized in the environment of the main process.
This also solves the warning `UserWarning: MathSAT convert(): UF with bool arguments have been translated warn("MathSAT convert(): UF with bool arguments have been translated")` which was caused by the missing contextualization of formulas in the workers' environments.